WIP Resizing Panels - Flux - Store
Started to implement the Store as a base class which can be inhareted from.

Found this really helpful artical - http://blog.andrewray.me/flux-for-stupid-people/

Take a look at the top comment by Ian Obermiller
Looks like I should go ahead and depend on a global variable to hold the dispatcher and store. It might could go in the app.jsx and then pass state down through all child props. But the code would be more flexible if none of the components were depending on their parents for state updates.