WIP Resizing Panels - Sort of fixed Horizontal
Ok so found a solution for horizontal that sort of works. Needs some love though. I'm thinking this needs to work in a flux fashion as well. Maybe containers keep a list of what they have and when something is resizing it reaches out to it's parent and then the parent manages the children...