Horizontal Resize works perfectly now!
Outstanding Edge Cases to account for:
1. Resizing something too big (pulling too far to the right) can cause items to it's left to start to shrink. Need some sort of min-width property so the resize event starts to ignore your changes, same goes for values less than 0.

2. Browser's cannot detect a mouse up event if the browser doesn't have focus. If resize is started and the mouse is dragged off screen and then released. Upon returning the resize event will continue. Might just need to accept this one. It's a rare edge-case and the user can simply click again to end the oporation.

Why the Queue:
If all cells within a table have a width expressed in px, the browser treats those values as a percentage and adjusts their display appropriately. The problem with having panels calculate their own width and update themselves was that the browser updates with each panel update. So as soon as one panel calculates it's width and sets state - the browser reflows the table. Making the calculations of the rest of the panels incorrect. This resulted in a slight pop whenever resizing a nested panel after it's parent container had been resized. To fix this, the parent container will now coordinate the effort. First collecting the accorite pixel widths in a queue, _then_ updated them all with the correct width value. Now it is safe to resize outer containers and then their inner panels without any vissible popping.