Remember width before reflow - also better auto width calculation
So ComponentDidMount ONLY gets called once after the first render. For all subsequent renders (Like if the container reflows itself virtically) we need to use the componentDidUpdate metthod.  Now even if no width was set, a panel's auto width will be replaced with a px value right after it's been rendered to the dom.

Best of all, the container will now remember all of it's child panel width values from the last resize event if it was in a horizontal state. Then whenever switching back to a horizontal state from a virtical state, it will reapply all the previously set widths, so your resized panels are exactly where you lefft them before collapsing into a virtical layout.

The addition of the purgeNulls method came about from an odd situation where the refs array would become poluted with undefined entries after switching to a virtical state and then back. Still not entirely sure why that was happening. Spent a lot of time tracking that down and eventually decided to go with the purge approach which is ammitedly more like a bandaid fix.