Vertical Resizing is sort of working
Ok so vertical resizing is sort of working but the calculating height keeps increasing for some reason. Not sure why since I'm dealing with nested containers vertically the same way I am horizontally. Also the problem where the resizing store isn't always updating the first time is persisting. Really not sure what could be causing that and if switching to browserify will fix it.