Resizing Panels is Solid!
The code needs some tidying up but I want to make a commit now while it's working. Figured out the two issues with vertical resizing.

There needs to only be one div in each panel's content. This won't be a problem in production with react components. But for me that extra span I had been appending with the width and height was throwing the calculations off.  So I've changed the empty panel's to just show a span with width and height there.

The other problem was with nested vertical containers getting an explicit height value. That was prevented outer vertical panels from being resized up because all of the nested vertical panels above had a hardcoded height. To fix this, the container now does a little magic in the resizeEvent started step.  First, all panels now have a new property 'active'. And the clicked panel will set that to true, and then back to false after the event is over.
It can't be in setState because of a bad race condition. Props might work just as well but oh well. This is less typing. Ok so with that active property on the panels, each container can check it's child panels to know if one of it's child panels is the one being resized. If not, it sets its bottom panel to 'auto'. This allows any panels below to be shifted up. Then after the event is over the updateChildPanelSizes code put's that bottom panels height value back for the next resize event.

In this commit I had moved all the height logic to Rows, but I don't think that was necessery. In my next commit I'm going to clean up the code and attempt to put height back in panels.