Use State for Active - Addressing race conditions
So sometimes setState doesn't happen syncronously.  This can be problematic for other modules needing to get the current state of a reference.  Before I had avoided this by mutating a property on the reactClass itself. However I just learned that setState can take a callback function. So I've corrected the code to use this better solution.