It's really holding more than just widths
First, we're going to use this lastChildPanelsInfo object to hold more than just widths so let's name it more appropriately.

Secondly, I've tried moving this to state and using setState to update it. For some reason that makes me lose all the refs? For now I'm leaving it where it is and focusing on getting the selectedToolIndex rememberence to work.