Tools out of sync bug solved
So it's really important to keep anything unique to an instance inside of it's state. Because everything else in a reactClass is shared with all instances. Before putting listener IDs on the class object itself was causing one unmount to remove the listeners for all instances. This new method is still scalable but also stable now. I'm not committing the same change on the container just yet because for some reason making that change on the container broke child panel refs from coming through. So that's another bug I need to deal with.