Squashing Bugs
Huge refactoring. Containers and panels now both get childPanel info from the UI_Store.
No more using refs and much better merging of old panel info with new panel info.

Removed dependency on micro-event.js
It was mainly there because it was cool but now the store works in a consitent way with the Dispatcher. And more importantly, stores are now able to emit event info so that subscribers can either be dumb and respond to everything or make a more informed decision. This is breaking the FLUX design a little but it's a substantial performance improvement having stores decide what to do with events and then allowing store subscribers to further decide what to do with events.