It's alive, IT'S ALIVE!
Finally squashed all those pesky bugs.  This thing is super stable now.

Still need to get containers to not use class properties for listeners. It's only working right now because no containers are being distroyed. But once merging and splitting are implemented that will break so the next thing I need to do is fix containers to manage StoreListeners in the same way as panels.