WIP - THIS NOT A WORKING COMMIT
LOTS of changes in this commit. Most notibly the move to the latest version of react. Which has broken a lot of the code.
This resource was really helpful in understanding the right way to do things,
https://www.youtube.com/watch?v=nQo0EdHNjto

Again, this is NOT a working commit. Just want to save what I've done before taking a breather for the weekend.