Resizing is nearly there
Added in the container to props. This is important because containers emit an action only after all of their child panels have been created. So it's the perfect time for all child panels to recalculate their width/height.

Removed the if condition, I was wrong about that. Basically for any event other than resizing we want to recalculate the width and height.