Resizing works perfectly again
So that 0.5 border width was throwing all the calculations off. Not to mention that Safari wouldn't render them at all. With that removed the panels accurately calculate their width/height and it seams that there is no race condition. I guess all the panels call set-state but the dom isn't updated until react finishes it's diffing algorithm. What I had thought was a race condition was actually my weird css

The downside is that now I've got to figure another way to show the borders without putting it on panels. Maybe containers would make more sense