Return current state not initial state
Found a bug in the reducers. State is either passed in or defaults to initial data in the function declarations. So in the switch cases the default actions need to be to return whatever state was passed to the function. Otherwise every action the switch case doesn't recognize will cause the store to reset to initial state.