Initiate Layout
The ReduxUIPanels component now recursively parses it's children and builds out the initial page layout in a single object in a single Redux action. With this the Containers and panels will be able to build from the store and we can remove Containers and Panels from the store. The register events won't be necessary for initial setup but something like them will need to be around for spliting and merging panels. Maybe just panelSplit, and panelMerge events? That solves that :)