Container is functional and reworking the layout object
Container is now a purely functional component - although not yet functional

Reworking the layout object to make looking up component props easier.

I now have each component aware of it's parent's id in the layout. This eliminate the need for a higherarchy. So next step will be to flatten it.