Panels now also build completely from the redux store
All a panel needs to know is it's type, id, and parentContainer flow - which the parentContainer can easily provide just by knowing it's own id, So now the Redux store is the source of truth for the entire layout.

Panels get a list of their children dynamacly generated from the store's flat index of everything in the ui. By removing all nesting it should be trivial to add splitting and merging panels. Just, add a key in the index of type Panel with a new id and pass in the right parent id. Next time everything renders it will have one recognize that new child and just work.

TODO - with all the changes resizing is temporarily broken. Need to fix that before getting to the fun stuff.